Move ServerProfileInfoService to data module
