+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/service/network/ServerProfileInfoService.kt
@@ -1,4 +1,4 @@
-package de.tum.informatics.www1.artemis.native_app.core.datastore
+package de.tum.informatics.www1.artemis.native_app.core.data.service.network
 
 import de.tum.informatics.www1.artemis.native_app.core.data.NetworkResponse
 import de.tum.informatics.www1.artemis.native_app.core.model.server_config.ProfileInfo
